// src/app/enums/column-type.enum.ts
export enum ColumnType {
    Currency = 1,
    Number = 2,
    Date = 3,
    Text = 4,
    Percent = 5,
    Download = 6,
    Component = 7,
    TextBox = 8,
}
export enum PolicyDataType {
    Currency = 1,
    Number = 2,
    Date = 3,
    Text = 4,
    Percent = 5,
    Download = 6,
    Component = 7,
    TextBox = 8
}